Implement Wallet, Operation and Diagram models
Wallet's main responsibility is to keep the amount of money that user
keeps. Each time the Income or Outcome is made on the Wallet the
notification is made to the Diagram to register that new operation has
been made.
Diagram's main responsibility is to store all the operations made on the
Wallet, so they can be published to the View. Diagram reacts on the
Wallet notification with the Update() method and adds new operation with
the latest amount of it.
Operation class represents single operation that is made on the Wallet.
It implements the Prototype design pattern so it can be easily created
based on the latest operation and it's data.