Cover all models with unit tests
The unit tests covers all models that are currently implemented.
There are two classes that tests three methods, but note that Chart and
Operation are strongly related
(Chart is based on Operation, and everything that Chart displays is a
result of Operation's class activity)
so there's only one class to test both these.

NOTE: Some minor changes were made to the models due to unit tests
failing. It turned out that implementation was not fully correct.