Design and create View with prototyped bindings
The simple View for the Plutus app contains one main window with three
sections:
Left panel for operations on the user's account,
Right panel for the chart to be displayed,
Botton panel where current amount of money in the Wallet will be displayed.

Bindings are created but they are made based on assumptions in the names
of ViewModels and their fields.