Improve ViewModel approach by introducing Facade
The Facade design pattern is responsible for managing all it's
subclasses when it comes to various logic and/or interfaces behind them.
In case of this application it was used to allow all the ViewModels to
be gathered in one place instead of constructing them and calling them
directly in the MainWindow code-behind class.

This commit also implements the basic ViewModels for both Chart and Wallet.