Minor corrections due to changes in ViewModel
Changes in the ViewModel implemented according to the requested behavior
forced the Models mechanism to act different. So the unit tests had to
be adjusted so there's no fail.
Also the View had one correction to be made in the commands attached to
the buttons: It turns out that Income command was attached to both
buttons.