Bitbucket -> GitHub repository migration (#1)
* Initial commit

README and gitignore has been added to the project.

* Add empty solution

The initial solution for Visual Studio contains two projects:
 - App; where application (frontend and backend) will be implemented
 - Tests; where all the unit testing will be implemented
The solution is automaticaly generated by VS IDE but it's prepared for
development (unused usings removed, spaces and indents adjusted).

* Put MainWindow class in the View directory

This commit reorganizes the directory structure into three separate
sections (directories) related to the MVVM pattern elements: View,
ViewModels, Models.
The View directory now contains the MainWindow class with it's
code-behind class.

NOTE: The csproj file contains also two other directories included in
the build even though they were not yet created. They will appear in the
commit once any file is placed in them.

* Minor corrections in the README document

The corrections include:
 - Last purpose now correctly pointed out
 - patterns and their desctiptions now presented in the table for
   comfortable reading

* Design and create View with prototyped bindings

The simple View for the Plutus app contains one main window with three
sections:
Left panel for operations on the user's account,
Right panel for the chart to be displayed,
Botton panel where current amount of money in the Wallet will be displayed.

Bindings are created but they are made based on assumptions in the names
of ViewModels and their fields.

* Correct App dependencies with OxyPlot and App class

The OxyPlot is a chart drawing package from NuGet used in this project
to draw the plot of user's operations and finances.
App class is renamed to MainApp due to confusion of auto-generating tool
when creating partial App class in App namespace.

* Implement Wallet, Operation and Diagram models

Wallet's main responsibility is to keep the amount of money that user
keeps. Each time the Income or Outcome is made on the Wallet the
notification is made to the Diagram to register that new operation has
been made.
Diagram's main responsibility is to store all the operations made on the
Wallet, so they can be published to the View. Diagram reacts on the
Wallet notification with the Update() method and adds new operation with
the latest amount of it.
Operation class represents single operation that is made on the Wallet.
It implements the Prototype design pattern so it can be easily created
based on the latest operation and it's data.

* Cover all models with unit tests

The unit tests covers all models that are currently implemented.
There are two classes that tests three methods, but note that Chart and
Operation are strongly related
(Chart is based on Operation, and everything that Chart displays is a
result of Operation's class activity)
so there's only one class to test both these.

NOTE: Some minor changes were made to the models due to unit tests
failing. It turned out that implementation was not fully correct.

* Improve ViewModel approach by introducing Facade

The Facade design pattern is responsible for managing all it's
subclasses when it comes to various logic and/or interfaces behind them.
In case of this application it was used to allow all the ViewModels to
be gathered in one place instead of constructing them and calling them
directly in the MainWindow code-behind class.

This commit also implements the basic ViewModels for both Chart and Wallet.

* Make a RelayCommand a base to implement all the commands

Each command of the application is based on the RelayCommand, because
it's mechanism is the same no matter what command from which ViewModel
is executed.

* Minor corrections due to changes in ViewModel

Changes in the ViewModel implemented according to the requested behavior
forced the Models mechanism to act different. So the unit tests had to
be adjusted so there's no fail.
Also the View had one correction to be made in the commands attached to
the buttons: It turns out that Income command was attached to both
buttons.

* Update Chart and draw lines for both operations and summary

* User

User represents the Singleton design pattern and it's used to keep
globally (but safely) all the data related to the user of this
application.

* Corrections and comments in the full implementation

* Fix: Unit tests build fails

* Update README

The information updated in README documents are:
 - Design patterns used: Decorator was switched to Facade, and Singleton
   is now used only for the User class session
 - Technology and tools table was extended with the OxyPlot from NuGet

* Add name of an User to the Operation's data

Each Operation keeps the information of itself. Now there's also the
name of an user who performed that operation.

* Improve plot by range of axis and their titles

* Add right panel with Operation's details

The new panel on the right of the application main window provides the
application with the functionality of displaying one selected Operation
element from the list.
The selected Operation element is displayed in the details presented on
this new panel and includes all of properties that Operation has.

* Refactor and improve the code

The improvements are mostly done to the code style, more comments
explaining the design patterns related sections.
Also several solutions were reimplemented.

* Replace .gitignore with more standarized

The new .gitignore file is the one used for all projects developed using
C++/C# technologies especially with Visual Studio.

* Include solution configuration in the version control

Previous approach ignored the solution configuration file, but the
current approach have it in the repository.

* Correct build output paths

There are two projects in the Plutus solution, each of them building
different type of project.
New approach keeps all the build results in the Build/ directory, so
this commit sets this as a common place to put all the results.